--- src/Desafio.Pokemon.Business/Models/CPF.cs
@@ -1,12 +1,0 @@
-﻿namespace Desafio.Pokemon.Business.Models
-{
-    public class CPF
-    {
-        public string Numero { get; set; }
-
-        public bool Validar()
-        {
-            return Numero.Length == 11;
-        }
-    }
-}
